# SwaggerOffline.yaml
openapi: 3.1.0
info:
  title: Auto Escola S164
  description: API Rest para gerenciamento de instruções da Auto Escola S164
  contact:
    name: Auto Escola S164
    email: autoescolas164@gmail.com.br
  license:
    name: Apache 2.0
    url: http://springdoc.org
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /usuarios:
    put:
      tags:
        - usuario-controller
      operationId: atualizarUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosAtualizacaoUsuario"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoUsuario"
    post:
      tags:
        - usuario-controller
      operationId: cadastrarUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosCadastroUsuario"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoUsuario"
  /instrutores:
    get:
      tags:
        - instrutor-controller
      operationId: listarInstrutores
      parameters:
        - name: paginacao
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/PageDadosListagemInstrutor"
      security:
        - bearer-key: [ ]
    put:
      tags:
        - instrutor-controller
      operationId: atualizarInstrutor
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosAtualizacaoInstrutor"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoInstrutor"
      security:
        - bearer-key: [ ]
    post:
      tags:
        - instrutor-controller
      operationId: cadastrarInstrutor
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosCadastroInstrutor"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoInstrutor"
      security:
        - bearer-key: [ ]
  /alunos:
    get:
      tags:
        - aluno-controller
      operationId: listarAlunos
      parameters:
        - name: paginacao
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/PageDadosListagemAluno"
    put:
      tags:
        - aluno-controller
      operationId: atualizarAluno
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosAtualizacaoAluno"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAluno"
    post:
      tags:
        - aluno-controller
      operationId: cadastrarAluno
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosCadastroAluno"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAluno"
  /login:
    post:
      tags:
        - login-controller
      operationId: efetuarLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosLogin"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosTokenJWT"
  /instrucoes:
    post:
      tags:
        - instrucao-controller
      operationId: agendarInstrucao
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosAgendamento"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAgendamento"
  /usuarios/{id}:
    get:
      tags:
        - usuario-controller
      operationId: detalharUsuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoUsuario"
  /instrutores/{id}:
    get:
      tags:
        - instrutor-controller
      operationId: detalharInstrutor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoInstrutor"
      security:
        - bearer-key: [ ]
    delete:
      tags:
        - instrutor-controller
      operationId: excluirInstrutor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
      security:
        - bearer-key: [ ]
  /healthy-check:
    get:
      tags:
        - healthy-check-controller
      operationId: healthyCheck
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: string
  /alunos/{id}:
    get:
      tags:
        - aluno-controller
      operationId: detalharAluno
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAluno"
    delete:
      tags:
        - aluno-controller
      operationId: excluirAluno
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
  /usuarios/id:
    delete:
      tags:
        - usuario-controller
      operationId: excluirUsuario
      parameters: [ ]
      responses:
        "200":
          description: OK
  /instrucoes/{id}:
    delete:
      tags:
        - instrucao-controller
      operationId: deletar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
components:
  schemas:
    DadosAtualizacaoUsuario:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        senha:
          type: string
        perfil:
          type: string
          enum:
            - USER
            - ADMIN
    DadosDetalhamentoUsuario:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        senha:
          type: string
        perfil:
          type: string
          enum:
            - USER
            - ADMIN
    DadosAtualizacaoInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
    DadosEndereco:
      type: object
      properties:
        cep:
          type: string
          minLength: 1
          pattern: "[0-9]{5}-[0-9]{3}"
        logradouro:
          type: string
          minLength: 1
        numero:
          type: string
        complemento:
          type: string
        bairro:
          type: string
          minLength: 1
        cidade:
          type: string
          minLength: 1
        uf:
          type: string
          minLength: 1
          pattern: "[A-Z]{2}"
      required:
        - bairro
        - cep
        - cidade
        - logradouro
        - uf
    DadosDetalhamentoInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        cnh:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        endereco:
          $ref: "#/components/schemas/Endereco"
        ativo:
          type: boolean
    Endereco:
      type: object
      properties:
        cep:
          type: string
        logradouro:
          type: string
        numero:
          type: string
        complemento:
          type: string
        bairro:
          type: string
        cidade:
          type: string
        uf:
          type: string
    DadosAtualizacaoAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
    DadosDetalhamentoAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        cpf:
          type: string
        email:
          type: string
        telefone:
          type: string
        endereco:
          $ref: "#/components/schemas/Endereco"
    DadosCadastroUsuario:
      type: object
      properties:
        login:
          type: string
          minLength: 1
        senha:
          type: string
          minLength: 1
        perfil:
          type: string
          enum:
            - USER
            - ADMIN
      required:
        - login
        - perfil
        - senha
    DadosLogin:
      type: object
      properties:
        login:
          type: string
          minLength: 1
        senha:
          type: string
          minLength: 1
      required:
        - login
        - senha
    DadosTokenJWT:
      type: object
      properties:
        tokenJWT:
          type: string
    DadosCadastroInstrutor:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        telefone:
          type: string
          minLength: 1
        cnh:
          type: string
          minLength: 1
          pattern: \d{11}
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
      required:
        - cnh
        - email
        - especialidade
        - nome
        - telefone
    DadosAgendamento:
      type: object
      properties:
        id_aluno:
          type: integer
          format: int64
        id_instrutor:
          type: integer
          format: int64
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        data_hora:
          type: string
          format: date-time
      required:
        - data_hora
        - id_aluno
    DadosDetalhamentoAgendamento:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome_aluno:
          type: string
        nome_instrutor:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        data_hora:
          type: string
          format: date-time
    DadosCadastroAluno:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        cpf:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        telefone:
          type: string
          minLength: 1
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
      required:
        - cpf
        - email
        - nome
        - telefone
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    DadosListagemInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
    PageDadosListagemInstrutor:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/DadosListagemInstrutor"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    DadosListagemAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
    PageDadosListagemAluno:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/DadosListagemAluno"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
  securitySchemes:
    bearer-key:
      type: http
      scheme: bearer
      bearerFormat: JWT